+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- option `--debug-cc1101`
 
 ## [0.2.0] - 2020-12-11
 ### Changed
@@ -14,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - option `--unlock-spi-device` to release the `flock` from the SPI device file
   after configuring the transceiver
 
-### [0.1.1] - 2020-12-08
+## [0.1.1] - 2020-12-08
 ### Fixed
 - syntax errors on python3.5